pyproject.toml
[project]
name = "linkedin-content-automator-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flasgger>=0.9.7.1",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "flask-sqlalchemy>=3.1.1",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.4",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-huggingface>=1.2.0",
    "transformers",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
